# Fix type instabilities in compactified methods